Fix import and structure
