+++ DNAnet/data/data_models/hid_dataset.py
@@ -16,7 +16,7 @@ from DNAnet.data.data_models import Panel
 from DNAnet.data.data_models.base import InMemoryDataset, SimpleDataset
 from DNAnet.data.data_models.hid_image import HIDImage, Ladder
 from DNAnet.data.data_models.structs import AlleleAnnotation, ScanpointAnnotation
-from DNAnet.data.strategies.dataset_strategies import DatasetStrategy
+from DNAnet.data.strategies.dataset_strategies import DatasetStrategy, NFI_RND_DatasetStrategy
 from DNAnet.data.strategies.kit_strategies.scaling_strategy import ScalingStrategy
 from DNAnet.data.strategies.sample_validation_strategy import SampleValidationStrategy
 from DNAnet.data.split import split_data_in_k_folds
@@ -25,8 +25,7 @@ from DNAnet.typing import PathLike
 from DNAnet.utils import (
     get_noc_from_rd_file_name,
     get_prefix_from_filename,
-    is_rd_hid_filename,
-    load_donor_alleles,
+    is_rd_hid_filename
 )
 
 
@@ -155,6 +154,7 @@ class HIDDataset(InMemoryDataset):
                 LOGGER.info(f"Writing data to arrow cache: {self.cache_path}")
                 write_to_hf_cache(self.cache_path, self._data, include_size_standard)
 
+        # TODO improve loading of annotations and ground truth annotations
         if self.ground_truth_as_annotations:
             LOGGER.info("Loading ground truth annotations")
             for image in self._data:
@@ -168,7 +168,7 @@ class HIDDataset(InMemoryDataset):
                             image.path.stem
                         )
                     else:
-                        image._meta["called_alleles"] = load_donor_alleles(image.path.stem, self._panel)
+                        image._meta["called_alleles"] = NFI_RND_DatasetStrategy.load_donor_alleles(image.path.stem, self._panel)
                 except ValueError as e:
                     LOGGER.warning("Could not load ground truth annotations for %s: %s", image.path, e)
 
@@ -586,12 +586,12 @@ class HIDDataset(InMemoryDataset):
             :param scaler: numpy array containing the scaler values to be used for finding the closest scanpoint indices.
             :return: ScanpointAnnotation object containing the translated scanpoint annotation.
         """
-        # TODO: do not hardcode amount of scanpoints
-        scanpoint_annotation = np.zeros((StrategyRegistry.get_kit().num_dyes, 4096), dtype=np.int8)
+        scanpoint_annotation = np.zeros((StrategyRegistry.get_scaling_strategy().kit.num_dyes,
+                                         StrategyRegistry.get_scaling_strategy().scanpoint_resolution), dtype=np.int8)
         for locus in allele_annotation.annotation:
             for allele in locus.alleles:
                 # for each allele, find the left and right bin of the allele using the panel that has been adjusted by the corresponding ladder.
-                bp, left_bin, right_bin = adjusted_panel.get_allele_basepair_and_bins(locus.name, allele.name)
+                _, left_bin, right_bin = adjusted_panel.get_allele_basepair_and_bins(locus.name, allele.name)
 
                 # use the scaler to find the closest scanpoint indices for the left and right bins of the allele
                 left_scanpoint = np.argmin(np.abs(scaler - left_bin))
