+++ DNAnet/data/parsing/parse_annotations.py
@@ -51,35 +51,6 @@ def parse_called_alleles(annotation_file_path: PathLike,
         return None
 
 
-def translate_allele_to_scanpoint_annotation(allele_annotation: AlleleAnnotation, adjusted_panel: Panel, scaler: np.ndarray) -> ScanpointAnnotation:
-    """
-    Translates allele annotation to scanpoint annotation by finding the closest scanpoint indices for the left and right bins
-    of each allele using the provided scaler and adjusted panel. The entire allelic bin is annotated as 1.
-
-    All non annotated scanpoints are labeled as 0, while annotated scanpoints are labeled as 1.
-    The resulting scanpoint annotation is a binary matrix of shape (num_dyes, num_scanpoints)
-
-        :param allele_annotation: AlleleAnnotation object containing the allele annotations to be translated.
-        :param adjusted_panel: Panel object containing the adjusted panel information to be used for translation.
-        :param scaler: numpy array containing the scaler values to be used for finding the closest scanpoint indices.
-        :return: ScanpointAnnotation object containing the translated scanpoint annotation.
-    """
-    scaling_strategy = StrategyRegistry.get_scaling_strategy()
-    scanpoint_annotation = np.zeros((scaling_strategy.kit.num_dyes, scaling_strategy.scanpoint_resolution), dtype=np.int8)
-    for locus in allele_annotation.annotation:
-        for allele in locus.alleles:
-            # for each allele, find the left and right bin of the allele using the panel that has been adjusted by the corresponding ladder.
-            bp, left_bin, right_bin = adjusted_panel.get_allele_basepair_and_bins(locus.name, allele.name)
-
-            # use the scaler to find the closest scanpoint indices for the left and right bins of the allele
-            left_scanpoint = np.argmin(np.abs(scaler - left_bin))
-            right_scanpoint = np.argmin(np.abs(scaler - right_bin))
-
-            scanpoint_annotation[locus.dye_row, left_scanpoint : right_scanpoint] = 1
-
-    return ScanpointAnnotation(annotation=scanpoint_annotation)
-
-
 
 def _parse_annotations(panel: Panel, results, allele_cols, height_cols) \
         -> List[Marker]:
